% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_topography.R
\name{elev.density.df}
\alias{elev.density.df}
\title{Convert an elevation raster to a data frame suitable for density plotting}
\usage{
elev.density.df(raster = NULL, raster.path = NULL, frame = NULL,
  frame.path = NULL, y.resolution = 1, n.y = NULL, x.resolution = 1,
  n.x = NULL, elev.scale = 1, na.sub = 0)
}
\arguments{
\item{raster}{Raster object. The elevation raster to use. If not provided, then a raster will be read in from \code{raster.path}. Defaults to \code{NULL}.}

\item{raster.path}{Character string. The full filepath to the elevation raster, including filename and file extension. Passed to \code{raster::raster()}. Only used if \code{raster} is \code{NULL}. Defaults to \code{NULL}.}

\item{frame}{Optional spatial polygons or spatial polygons data frame. A frame to crop the elevation raster to. If \code{NULL}, then \code{frame.path} will be used. If \code{frame.path} is also \code{NULL} then no cropping will be applied. Defaults to \code{NULL}.}

\item{frame.path}{Optional character string. The full filepath to the frame shapefile, including filename and file extension (".shp"). Only used if \code{frame} is \code{NULL}. If both are \code{NULL} then the raster will not be cropped to a frame. Defaults to \code{NULL}.}

\item{y.resolution}{Numeric value. The effective resolution on the y axis, e.g. \code{0.1} will result in keeping 1/10th of the values whereas \code{1} will keep all of them. Smaller values can dramatically increase speed. Defaults to \code{1}.}

\item{n.y}{Optional numeric value. An integer number of y values to keep; determines the resolution of the plot on the y axis and the number of ridgelines (if plotting with \code{geom_density_ridges()}). This is applied after \code{y.resolution}. Will be rounded to the nearest integer and due to other rounding issues may result in \code{n.y +/- 1} unique y values. Defaults to \code{NULL}.}

\item{x.resolution}{Numeric value. The effective resolution on the x axis, e.g. \code{0.1} will result in keeping 1/10th of the values whereas \code{1} will keep all of them. Smaller values can dramatically increase speed. Defaults to \code{1}.}

\item{n.x}{Optional numeric value. An integer number of x values to keep; determines the resolution of the plot on the x axis. This is applied after \code{x.resolution}. Will be rounded to the nearest integer and due to other rounding issues may result in \code{n.x +/- 1} unique x values. Defaults to \code{NULL}.}

\item{elev.scale}{Optional numeric value. All elevation values are multiplied by \code{elev.scale}, which, when between 0 and 1 can increase speed by reducing the number of rows of in the data frame. This will have a flattening effect on the ridges in the output plot. Defaults to \code{1}.}

\item{na.sub}{Optional numeric value. An integer to replace any NA values in the elevation data with. Defaults to \code{0}.}

\item{frame.filename}{Character string. The filename of the shapefile (without file extension) to be read from \code{frame.path}.}
}
\value{
Data frame with variables x and y, each observation corresponding to a cell in the original raster. Coordinates in x and y are relative to raster extent, not CRS, and observations are repeated in proporition to their relative cell value.
}
\description{
Creates a long/tall data frame from an elevation raster. Each coordinate pair is repeated the according to its relative elevation within the raster, e.g. if the minimum elevation is 100 and the maximum is 200, coordinates with an elevation of 100 will be repeated once and those with an elevation of 200 will be repeated 100 times.
}
