% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_topography.R
\name{kernel_scan}
\alias{kernel_scan}
\title{Return values from a kernel applied to a matrix}
\usage{
kernel_scan(
  matrix,
  search_distance_x = 50,
  search_distance_y = 20,
  search_value_min = 10,
  search_value_max = Inf,
  return_value = 1.5
)
}
\description{
Specify the dimensions of a rectangular kernel to scan the given matrix with
}
