% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_topography.R
\name{matrix_to_dataframe}
\alias{matrix_to_dataframe}
\title{Convert a matrix to a data frame}
\usage{
matrix_to_dataframe(matrix)
}
\arguments{
\item{matrix}{The matrix of elevation values to be converted.}
}
\value{
A data frame with the variables "x", "y", and "elev".
}
\description{
Convert a matrix of elevation data where columns represent the x component and rows represent the y to a data frame.
}
