% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_topography.R
\name{raster_to_dataframe}
\alias{raster_to_dataframe}
\title{Convert an elevation raster to a data frame}
\usage{
raster_to_dataframe(elev_raster)
}
\arguments{
\item{elev_raster}{The elevation raster to convert.}
}
\value{
A data frame with the variables "x", "y", and "elev".
}
\description{
Convert a raster where cell values correspond to elevation to a data frame where each cell is an observation with the variables "x", "y", and "elev".
}
