% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_topography.R
\name{ridgemap}
\alias{ridgemap}
\title{Create a topographic ridgeline map from elevation data}
\usage{
ridgemap(
  elev_data,
  line_color = "white",
  background_color = "gray10",
  line_count = NULL,
  y_scalar = 100,
  scale_factor = 1,
  line_weight = 0.5,
  min_height = 0
)
}
\arguments{
\item{elev_data}{Raster, matrix, or data frame. The elevation data. If a raster, the value in each cell must represent the elevation. If a matrix, the values must be the elevations and the rows and columns must correspond to the x and y coordinates, respectively. If a data frame, there must be the variables \code{"x"}, \code{"y"}, and \code{"elev"} containing the x coordinates, y coordinates, and elevations.}

\item{line_color}{Character string. The color for the ridgelines. Must be recognizable by \code{ggplot}, e.g. \code{"white"} or \code{"#ffffff"}. Defaults to \code{"white"}.}

\item{background_color}{Character string. The color for the ridgelines. Must be recognizable by \code{ggplot}, e.g. \code{"black"} or \code{"#000000"}. Defaults to \code{"gray10"}.}

\item{line_count}{Optional numeric. The number of ridgelines to draw. If this is \code{NULL} or greater than or equal to the number of available latitude lines in the data, all possible lines will be drawn. If it is fewer than the available latitudes, only this many lines will be drawn, using an evenly-spaced subset of the data. Defaults to \code{NULL}.}

\item{y_scalar}{Numeric. The scalar to multiply all y values by. Larger values can help if your ridgelines are squashed at the bottom of your map. Defaults to \code{75}.}

\item{scale_factor}{Numeric. The vertical scaling factor to apply to the elevations in the plot. Larger values result in more exaggerated topography. Defaults to \code{1}.}

\item{line_weight}{Numeric. The thickness of the ridgelines. Used as the \code{size} argument in the \code{geom_ridges()} call. Defaults to \code{0.5}.}

\item{min_height.}{Numeric. The minimum height to display. Used by \code{geom_ridges()}. Defaults to \code{0}.}
}
\value{
A ridgeline elevation map as a ggplot object
}
\description{
Using elevation data in the form of a raster, matrix, or data frame, create a map using ridgeline plotting.
}
