% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_topography.R
\name{topo.ridges.plot}
\alias{topo.ridges.plot}
\title{Create a ridgeline plot from a data frame of relative coordinates}
\usage{
topo.ridges.plot(data, x.variable = "x", y.variable = "y", scale = 1,
  size = 1, fill = "light blue", line.gradient.start = "white",
  line.gradient.end = "white", invert.y = FALSE, invert.x = FALSE)
}
\arguments{
\item{data}{Data frame. Must contain variables corresponding to relative coordinates of elevation data. Each observation of a coordinate pair must appear in the data frame a number of times equal to its relative elevation, e.g. a coordinate pair that is twice as high as another should occur two times as many times in the data frame.}

\item{scale}{Numeric value. Determines the vertical scaling of ridges; passed to \code{ggridges::geom_density_ridges(ggplot2::aes(scale))}. Defaults to \code{1}.}

\item{size}{Numeric value. Determines the weight of ridgelines; passed to \code{ggridges::geom_density_ridges(ggplot2::aes(size))}. Defaults to \code{1}.}

\item{fill}{Character string. The color for the background of the plot. Must be a color that ggplot can recognize (either the color name or hexadecimal); passed to \code{geom_density_ridges(fill)} and \code{theme(panel.background = element_rect(fill))}. Defaults to \code{"light blue"}.}

\item{line.gradient.start}{Character string. The color for the foreground ridgeline of the plot. Used as the start of an auto-generated color gradient through to the backmost ridgeline. Must be a color that ggplot can recognize (either the color name or hexadecimal); passed to \code{scale_color_continuous(low)}. Defaults to \code{"white"}.}

\item{line.gradient.end}{Character string. The color for the backmost ridgeline of the plot. Used as the end of an auto-generated color gradient through to the backmost ridgeline. Must be a color that ggplot can recognize (either the color name or hexadecimal); passed to \code{scale_color_continuous(high)}. Defaults to \code{"white"}.}

\item{invert.y}{Logical. If \code{TRUE} then all y values will by multiplied by -1. Defaults to \code{FALSE}.}

\item{invert.x}{Logical. If \code{TRUE} then all x values will by multiplied by -1. Defaults to \code{FALSE}.}
}
\value{
A ggplot object.
}
\description{
Takes a data frame of relative coordinates, repeated according to relative elevation, and creates a ridgeline plot. There will be one density ridgeline for each unique y axis value.
}
