% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_topography.R
\name{topo.ridges}
\alias{topo.ridges}
\title{Create a ridgeline plot from an elevation raster}
\usage{
topo.ridges(raster = NULL, raster.path = NULL, frame = NULL,
  frame.path = NULL, n.ridges = 50, x.resolution = 1, n.x = NULL,
  elev.scale = 0.1, na.sub = 0, scale = 15, size = 1,
  fill = "light blue", line.gradient.start = "white",
  line.gradient.end = "white", invert.y = TRUE)
}
\arguments{
\item{raster}{Raster object. The elevation raster to use. If not provided, then a raster will be read in from \code{raster.path}. Defaults to \code{NULL}.}

\item{raster.path}{Character string. The full filepath to the elevation raster, including filename and file extension. Passed to \code{raster::raster()}. Only used if \code{raster} is \code{NULL}. Defaults to \code{NULL}.}

\item{frame}{Optional spatial polygons or spatial polygons data frame. A frame to crop the elevation raster to. If \code{NULL}, then \code{frame.path} will be used. If \code{frame.path} is also \code{NULL} then no cropping will be applied. Defaults to \code{NULL}.}

\item{frame.path}{Optional character string. The full filepath to the frame shapefile, including filename and file extension (".shp"). Only used if \code{frame} is \code{NULL}. If both are \code{NULL} then the raster will not be cropped to a frame. Defaults to \code{NULL}.}

\item{n.ridges}{Optional numeric value. An integer number of y values to keep; determines the resolution of the plot on the y axis and the number of ridgelines (if plotting with \code{geom_density_ridges()}). Will be rounded to the nearest integer. Defaults to \code{50}.}

\item{x.resolution}{Numeric value. The effective resolution on the x axis, e.g. \code{0.1} will result in keeping 1/10th of the values whereas \code{1} will keep all of them. Smaller values can dramatically increase speed. Defaults to \code{1}.}

\item{n.x}{Optional numeric value. An integer number of x values to keep; determines the resolution of the plot on the x axis. This is applied after \code{x.resolution}. Will be rounded to the nearest integer. Defaults to \code{NULL}.}

\item{elev.scale}{Optional numeric value. All elevation values are multiplied by \code{elev.scale}, which, when between 0 and 1 can increase speed by reducing the number of rows of in the data frame. This will have a flattening effect on the ridges in the output plot. Defaults to \code{0.1}.}

\item{na.sub}{Optional numeric value. An integer to replace any NA values in the elevation data with. Defaults to \code{0}.}

\item{scale}{Numeric value. Determines the vertical scaling of ridges; passed to \code{ggridges::geom_density_ridges(ggplot2::aes(scale))}. Defaults to \code{15}.}

\item{size}{Numeric value. Determines the weight of ridgelines; passed to \code{ggridges::geom_density_ridges(ggplot2::aes(size))}. Defaults to \code{1}.}

\item{fill}{Character string. The color for the background of the plot. Must be a color that ggplot can recognize (either the color name or hexadecimal); passed to \code{geom_density_ridges(fill)} and \code{theme(panel.background = element_rect(fill))}. Defaults to \code{"light blue"}.}

\item{line.gradient.start}{Character string. The color for the foreground ridgeline of the plot. Used as the start of an auto-generated color gradient through to the backmost ridgeline. Must be a color that ggplot can recognize (either the color name or hexadecimal); passed to \code{scale_color_continuous(low)}. Defaults to \code{"white"}.}

\item{line.gradient.end}{Character string. The color for the backmost ridgeline of the plot. Used as the end of an auto-generated color gradient through to the backmost ridgeline. Must be a color that ggplot can recognize (either the color name or hexadecimal); passed to \code{scale_color_continuous(high)}. Defaults to \code{"white"}.}

\item{invert.y}{Logical. If \code{TRUE} then all y values will by multiplied by -1. Defaults to \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
A wrapper for elev.density.df() and topo.ridges.plot(). Converts an elevation raster to a topographic visualization using density ridgelines.
}
